adjust for save config type long and int

+++ composeApp/src/androidMain/kotlin/com/came.parkare/dashboardapp/infrastructure/source/remote/apiserver/ApiServerRepositoryImpl.kt
@@ -106,7 +106,7 @@ class ApiServerRepositoryImpl(
         )
     }
 
-    override suspend fun saveScreenConfigType(type: Int): Boolean {
+    override suspend fun saveScreenConfigType(type: Long): Boolean {
         preferences.put(CONFIG_TYPE, type)
         return true
     }
